+++ Assets/Resources/Prefabs/AkimichiPhotonManager.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &2733857022725649675
+--- !u!1 &1585072811369884453
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,22 +8,22 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 4103844606767581170}
-  - component: {fileID: 5801726556809273598}
+  - component: {fileID: 6478752164039561029}
+  - component: {fileID: 2140714015081443516}
   m_Layer: 0
-  m_Name: NetworkManager
+  m_Name: AkimichiPhotonManager
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &4103844606767581170
+--- !u!4 &6478752164039561029
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2733857022725649675}
+  m_GameObject: {fileID: 1585072811369884453}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
@@ -32,15 +32,15 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &5801726556809273598
+--- !u!114 &2140714015081443516
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2733857022725649675}
+  m_GameObject: {fileID: 1585072811369884453}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e2fd9fa17fd93d041a4d01838b7ed532, type: 3}
+  m_Script: {fileID: 11500000, guid: 1f91befd907511d489a43d3f5c052635, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
