+++ Assets/Resources/Prefabs/AkimichiPhotonManager.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c77c6c2c0556b21498ed1cec0ed1c130
+guid: b10a99a68833cd7429fbf11fca6a8819
 PrefabImporter:
   externalObjects: {}
   userData:
