+++ Assets/Scripts/Network/AkimichiPhotonManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1f91befd907511d489a43d3f5c052635
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
