+++ Assets/Scripts/Scenes/Home.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 3f3cdc082bbd49949bc99d61f286ffb9
+guid: fcc42d59bc1451f4eb9f26a4f4a382c3
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
