+++ Assets/Scripts/Scenes/Home/HomeScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e46f6bbacb6f7044ebbc775e45094c6b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
